+++ sources/app.cpp
@@ -205,5 +205,6 @@ extern void app_render(GLFWwindow* window, int frame_width, int frame_height)
 
 extern void app_render_3rdparty_libs()
 {
+    ImGui::TextURL("Wallet Cloud Services", nullptr, STRING_CONST("https://wallet.wiimag.com"));
     ImGui::TextURL("EOD Historical Data API", nullptr, STRING_CONST("https://eodhistoricaldata.com/r/?ref=PF9TZC2T"));
 }